testservice(person): create file person.service, modification of de p…

+++ src/person/person.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Person, Role } from './person.entity';
@@ -54,11 +54,40 @@ export class PersonService {
      * });
      * ```
      */
-    async create(createPersonDto: CreatePersonDto): Promise<Person>{
-        const newPerson = this.personRepository.create(createPersonDto);
-        // El save permite que se active el @BeforeInsert antes de que un usuario se cree, activando el hashing
-        return this.personRepository.save(newPerson);
-    }
+    async create(createPersonDto: CreatePersonDto): Promise<{
+        
+        message: string;
+        statusCode: number;
+        data: Person;
+    }>{
+        try {
+            // person exists
+            const existingPerson = await this.personRepository.findOne({ where: { email: createPersonDto.email } });
+            if (existingPerson) {
+                return {
+                    message: 'Person already exists',
+                    statusCode: HttpStatus.CONFLICT,
+                    data: existingPerson
+                };
+            }
+            const newPerson = this.personRepository.create(createPersonDto);
+            // El save permite que se active el @BeforeInsert antes de que un usuario se cree, activando el hashing
+            const saved = await this.personRepository.save(newPerson);
+            return {
+                message: 'Person created successfully',
+                statusCode: HttpStatus.CREATED,
+                data: saved
+            };
+            } catch (error) {
+                throw new HttpException(
+                    {
+                        message: 'Error creating person',
+                            error: error.message,
+                        },
+                        HttpStatus.BAD_REQUEST,
+                    );
+                    }
+        }
 
     /**
      * Busca una persona por su correo electrónico
@@ -81,13 +110,33 @@ export class PersonService {
      * const personWithPass = await personService.findOneByEmail("juan@example.com", true);
      * ```
      */
-    async findOneByEmail(email: string, includePassword = false): Promise<Person | null> {
-        const findOptions = {
-            where: { email },
-            select: includePassword ? ['id', 'name', 'email', 'role', 'password'] : ['id', 'name', 'email', 'role'],
-        } as any;
+    async findOneByEmail(email: string, includePassword = false): Promise<
+        { 
+            message: string;
+            statusCode: number;
+            data: Person | null;
+        }> {
         
-        return this.personRepository.findOne(findOptions);
+        try {
+            const findOptions = {
+                where: { email },
+                select: includePassword ? ['id', 'name', 'email', 'role', 'password'] : ['id', 'name', 'email', 'role'],
+            } as any;
+        
+        return {
+            message: 'Person found successfully',
+            statusCode: HttpStatus.OK,
+            data: await this.personRepository.findOne(findOptions)
+        };
+        } catch (error) {
+            throw new HttpException(
+                {
+                    message: 'Error finding person by email',
+                    error: error.message,
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        }
     }
 
     /**
@@ -100,8 +149,27 @@ export class PersonService {
      * const allPersons = await personService.findAll();
      * ```
      */
-    findAll() {
-        return this.personRepository.find();
+    async findAll(): Promise <{
+        message: string;
+        statusCode: number;
+        data: Person[];
+    }> {
+        try {
+            const allPersons = await this.personRepository.find();
+            return {
+                message: 'All persons retrieved successfully',
+                statusCode: HttpStatus.OK,
+                data: allPersons
+            };
+        } catch (error) {
+            throw new HttpException(
+                {
+                    message: 'Error retrieving all persons',
+                    error: error.message,
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        }
     }
 
     /**
@@ -120,9 +188,26 @@ export class PersonService {
      * const patients = await personService.findByrole(Role.Patient);
      * ```
      */
-    findByrole(role: Role) {
-        return this.personRepository.find({ where: { role } });
-    }
+    async findByrole(role: Role): Promise<{
+        message: string;
+        statusCode: number;
+        data: Person[];
+        }> {
+            try {
+                const persons = await this.personRepository.find({ where: { role } });
+                return {
+                    message: 'Persons found successfully',
+                    statusCode: HttpStatus.OK,
+                    data: persons
+                };
+            } catch (error) {
+                throw new HttpException(
+                    {message: 'Error finding persons by role', error: error.message},
+                    HttpStatus.BAD_REQUEST,
+                );
+                }
+        }
+
 
     /**
      * Busca una persona por su ID
@@ -135,8 +220,27 @@ export class PersonService {
      * const person = await personService.findOne(1);
      * ```
      */
-    findOne(id: number) {
-        return this.personRepository.findOne({ where: { id } });
+    async findOne(id: number): Promise<{
+        message: string;
+        statusCode: number;
+        data: Person | null;
+        }> {
+        try {
+            const person = await this.personRepository.findOne({ where: { id } });
+            return {
+                message: 'Person found successfully',
+                statusCode: HttpStatus.OK,
+                data: person
+            };
+        } catch (error) {
+            throw new HttpException(
+                {
+                    message: 'Error finding person by id',
+                    error: error.message,
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        }
     }
 
     /**
@@ -158,9 +262,37 @@ export class PersonService {
      * });
      * ```
      */
-    async update(id: number, updatePersonDto: UpdatePersonDto) {
-        await this.personRepository.update(id, updatePersonDto);
-        return this.personRepository.findOne({ where: { id } });
+    async update(id: number, updatePersonDto: UpdatePersonDto): Promise<{
+        message: string;
+        statusCode: number;
+        data: Person;
+    }> {
+        const person = await this.personRepository.findOne({ where: { id } });
+        if (!person) {
+            throw new HttpException(
+                {
+                    message: 'Person not found',
+                    error: 'Person not found',
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        } Object.assign(person, updatePersonDto);
+        try {
+            const updated = await this.personRepository.save(person);
+        return {
+            message: 'Person updated successfully',
+            statusCode: HttpStatus.OK,
+            data: updated
+        };
+        } catch (error) {
+            throw new HttpException(
+                {
+                    message: 'Error updating person',
+                    error: error.message,
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        } 
     }
     /**
      * Elimina una persona por su ID
@@ -173,9 +305,37 @@ export class PersonService {
      * await personService.remove(1);
      * ```
      */
-    remove(id: number) {
     // Delete person by id
-    async remove(id: number) {
-        return this.personRepository.delete(id);
+    async remove(id: number): Promise<{
+        message: string;
+        statusCode: number;
+        data: any;
+    }> {
+        const person = await this.personRepository.findOne({ where: { id } });
+        if (!person) {
+            throw new HttpException(
+                {
+                    message: 'Person not found',
+                    error: 'Person not found',
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        }
+        try {
+            const deleted = await this.personRepository.delete(id);
+            return {
+                message: 'Person deleted successfully',
+                statusCode: HttpStatus.OK,
+                data: deleted
+            };
+        } catch (error) {
+            throw new HttpException(
+                {
+                    message: 'Error deleting person',
+                    error: error.message,
+                },
+                HttpStatus.BAD_REQUEST,
+            );
+        }
     }
 }
